Model Refactor: Fix lazy loading
